feat: Adds scale and expanded scale variables.
Right now, developers have to multiple `$baseline` by integers to get values along the scale. Although this is trivia, it does introduce a source of divergence from the design system should someone forget what the scale values are. This commit introduces expanded variables for the scale and the values.

Not the that while the `json`, `custom-properties`, and 'scss' module formats the scale itself correctly, the `common.js` formatter does not. I'm not quite sure why, but I'm also not certain it's necessary in that format. We can revisit that if necessary though.